backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
paho-mqtt>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
joblib>=1.3.0
google-genai>=1.0.0
